move get_subject_metadata_from_rat_info_folder utility method to be accessible from other projects
