+++ src/constantinople_lab_to_nwb/utils/__init__.py
@@ -1,0 +1,1 @@
+from .get_subject_metadata import get_subject_metadata_from_rat_info_folder
